+++ core/src/test/resources/testApplicationContext.xml
@@ -12,9 +12,9 @@
 	<bean name="ibisManager" class="org.frankframework.testutil.mock.MockIbisManager" />
 
 	<bean id="MessageEventListener" class="org.frankframework.lifecycle.events.MessageEventListener" scope="singleton" />
-	<bean id="jndiContextPrefixFactory" class="org.frankframework.core.JndiContextPrefixFactory" />
+	<bean id="jndiContextPrefixFactory" class="org.frankframework.jndi.JndiContextPrefixFactory" />
 
-	<bean id="ResourceObjectLocator" class="org.frankframework.jdbc.factory.ResourceObjectLocator" />
+	<bean id="ResourceObjectLocator" class="org.frankframework.resourcelocator.ResourceObjectLocator" />
 	<bean id="JndiObjectLocator" class="org.frankframework.jndi.JndiObjectLocator" />
 	<bean id="dataSourceFactory" class="org.frankframework.testutil.TestDataSourceFactory" autowire="byType" primary="true"/>
