+++ core/src/test/resources/testConfigurationContext.xml
@@ -31,5 +31,5 @@
 
 	<!-- Required for the fake `lookupDestination` in the JmsMessagingSource -->
 	<bean id="jndiContextPrefix" factory-bean="jndiContextPrefixFactory" factory-method="getContextPrefix" />
-	<bean id="jndiContextPrefixFactory" class="org.frankframework.core.JndiContextPrefixFactory" />
+	<bean id="jndiContextPrefixFactory" class="org.frankframework.jndi.JndiContextPrefixFactory" />
 </beans>
