+++ core/src/test/resources/testDatabaseEnabledConfigurationContext.xml
@@ -19,7 +19,7 @@
 
 	<import resource="testConfigurationContext.xml"/>
 
-	<bean id="jndiContextPrefixFactory" class="org.frankframework.core.JndiContextPrefixFactory" />
+	<bean id="jndiContextPrefixFactory" class="org.frankframework.jndi.JndiContextPrefixFactory" />
 	<bean id="dataSourceFactory" class="${TestDataSourceFactory}" autowire="byType" primary="true"/>
 	<bean id="nonTransactionalDataSourceFactory" class="org.frankframework.jdbc.factory.NonTransactionalDataSourceFactory" autowire="byType" lazy-init="true"/>
 	<bean name="dbmsSupportFactory" class="org.frankframework.dbms.DbmsSupportFactory" />
