+++ ladybug/debugger/src/main/resources/springIbisTestTool.xml
@@ -93,10 +93,10 @@
 			<constructor-arg value="${ladybug.jdbc.datasource:}"/>
 		</bean>
 
-		<bean id="jndiContextPrefixFactory" class="org.frankframework.core.JndiContextPrefixFactory"/>
+		<bean id="jndiContextPrefixFactory" class="org.frankframework.jndi.JndiContextPrefixFactory"/>
 
 		<!-- Find DataSources in resources.yml and JNDI -->
-		<bean id="resourceObjectLocator" class="org.frankframework.jdbc.factory.ResourceObjectLocator">
+		<bean id="resourceObjectLocator" class="org.frankframework.resourcelocator.ResourceObjectLocator">
 			<property name="resourceFile" value="${ladybug.jdbc.resourceFile:resources.yml}"/>
 		</bean>
 		<bean id="jndiObjectLocator" class="org.frankframework.jndi.JndiObjectLocator" />
