+++ ladybug/debugger/src/main/resources/springIbisTestToolTibet2.xml
@@ -102,9 +102,9 @@
 	</bean>
 
 	<!-- START: These beans are no longer automatically present, and are only there when the database profile is used. -->
-	<bean id="jndiContextPrefixFactory" class="org.frankframework.core.JndiContextPrefixFactory"/>
+	<bean id="jndiContextPrefixFactory" class="org.frankframework.jndi.JndiContextPrefixFactory"/>
 
-	<bean id="resourceObjectLocator" class="org.frankframework.jdbc.factory.ResourceObjectLocator">
+	<bean id="resourceObjectLocator" class="org.frankframework.resourcelocator.ResourceObjectLocator">
 		<property name="resourceFile" value="${ladybug.jdbc.resourceFile:resources.yml}"/>
 	</bean>
 	<bean id="jndiObjectLocator" class="org.frankframework.jndi.JndiObjectLocator" />
